refactor(schema): reorganise into folder
