+++ server/api/gpt-typing-game.get.ts
@@ -1,6 +1,9 @@
 import OpenAI from 'openai';
+import type { z } from 'zod';
 import type { FetchError } from 'ofetch';
-import type { Character } from '~/types';
+import type { CharacterSchema } from '~/schemas/typing-game';
+
+type Character = z.infer<typeof CharacterSchema>;
 
 const config = useRuntimeConfig();
